Found a bug in constrain_bp and constrain_ss_bp -- they weren't creating the correct space. when trying to form the ss+bp space the tangent was not orthogonal to the dgrad and dvec. Still need to iron out convergence but overall decent.
Trying to do ic_reparam, only adding a new node if the icoords[nnR-1].gradrms or icoords[nnmax-nnP-1].gradrms < gaddmax and
dE[wstate2-2] < 5. kcal/mol. Exit criteria is that the totalgrad is < 0.05 and all the nodes are converged to 1 kcal
within the seam.

The optimizer is set to exit optimizing if either the gradient is really low and the dE is relatively low. Or if
the gradient is relatively low and the dE is really low.

Doing 11 nodes instead of 9